Trajectory time fixes (#43)
* fixing unrealistic polygon mirror deflectors:
now uses maxScanAngle_deg for estimation of angular space between pulses
now uses min(scanAngle_deg, maxEffectiveScanAngle_deg) for limiting pulses (anything above these values is discarded)

* adapted parameter scanAngleMax_deg for different rotating polygon mirror scanners

* gpstime is now calculated once when starting and then increased in "simulation time", i.e. using the PRF of the scanner

* GPSTime is now calculated properly (lead to overflow when using milliseconds in long)
GPSTime is now seconds since start of the GPS Week (Sunday, midnight)

* resetting GPS seconds if overflow in week is encountered (currentGpsTime_ms > 604800000.)

+++ src/scanner/detector/AbstractDetector.h
@@ -135,7 +135,7 @@ public:
 	    glm::dvec3 absoluteBeamOrigin,
 	    Rotation absoluteBeamAttitude,
 	    int state_currentPulseNumber,
-	    long currentGpsTime
+        double currentGpsTime
     ) = 0;
 
 	// ***  GETTERS and SETTERS  *** //
