+++ src/scanner/detector/FullWaveformPulseRunnable.h
@@ -96,7 +96,7 @@ public:
         glm::dvec3 absoluteBeamOrigin,
         Rotation absoluteBeamAttitude,
         int currentPulseNum,
-        long currentGpsTime,
+        double currentGpsTime,
         bool writeWaveform,
         bool calcEchowidth,
         std::vector<Measurement> * allMeasurements,
@@ -308,7 +308,7 @@ private:
 	    double max_time,
 	    glm::dvec3 & beamOrigin,
 	    glm::dvec3 & beamDir,
-	    long gpstime,
+        double gpstime,
 	    bool fullWaveNoise,
 	    RandomnessGenerator<double> &rg2
     );
